:- bug fix attempt with LIB._getAmountsForInitLP (market LP ticket token supply is too little) 	- now converting from 6 decimals to 18 decimals (for CallitTicket created) 		LP created was too little supply 	- refactored CONF.RATIO_LP_TOK_PER_USD from uint16 to uint32 		- updated required dependencies 			CONF.KEEPER_setLpSettings 			VAULT.createDexLP 			IVAULT.createDexLP 			ICONFIG.CONF.RATIO_LP_TOK_PER_USD NEXT: 	- required re-deploys (all still compile just fine) 		lib 		config, 		vault, 		delegate (for interface updates, uses CONF.RATIO_LP_TOK_PER_USD & invokes VAULT.createDexLP) 	- then test again from ‘makeNewMarket’ 		and then analyze supplies created for pair 		may have to perform a ADMIN_createPromoTicket & then buyTicketWithPromo 		 on one of the new tickets and check dexscreener for easier analysis
#================================================#
DEPLOYMENT algorithm… (full system & single contract updates)
	- deploy all contracts | single contract update (log addies created)
	- invoke CONF.KEEPER_setContracts (w/ input: addies created | 0x0 to opt-out)
		- invokes CONF_setConfig on all contracts (including opt-out/old addies as well)
			NOTE: only CONF contract can call these functions (unless “FIRST_”)

NEXT:
	- need more requirements for market maker earning $CALL
		ex: maker could create $100 LP and not promote,
			then declare himself winner, get his $100 back and earn free $CALL
	- test factory main feature sets via CLI and smartcontractgui.xyz
		DONE - deposit / send PLS
		DONE - FACT.makeNewMarket
		DONE - FACT.setMarketInfo
		DONE - FACT.getMarketForTicket
		DONE - FACT.getMarketForHash
		DONE - FACT.getMarketsForMakerOrCategory
		DONE - FACT.getMarketCntForMakerOrCategory
		DONE - FACT.getMarketHashesForMakerOrCategory
		DONE - DELEGATE.ADMIN_initPromoForWallet
		DONE - FACT.getPromosForAcct
		DONE - FACT.buyCallTicketWithPromoCode (requires DELEGATE.ADMIN_initPromoForWallet)
		- FACT.exeArbPriceParityForTicket (requires setting/forcing DEX environment)
		- FACT.closeMarketCallsForTicket (before/after ‘dtCallDeadline’)
		- FACT.initMyVoterHash & FACT.getMyVoterHash
		- FACT.castVoteForMarketTicket (before/after ‘dtResultVoteStart’ & ‘dtResultVoteEnd’)
		- FACT.closeMarketForTicket (before/after ‘dtResultVoteEnd’)
		- FACT.claimTicketRewards (before/after ‘dtResultVoteEnd’ & live|!live)
		- FACT.claimVoterRewards (w/wo active votes in ‘mark.live’ status, ensure no dup payments)
		- DELEGATE.claimPromotorRewards
	… … … … … …
	- review requirements between dt deadlines when making a new market
		- maybe need admin guaranteed time between each deadline (ie. at least 1 day to vote)
	- integrate stop for new markets created during manual contract updates
		- problem was w/ existing old markets & tickets (but is this still a problem w/ new CONFIG|M model?)
	- integrate function or some way to transfer over all variable settings during CONFIG migration
	- search for & review all ‘TODO’
	- last ‘LEFT OFF HERE’ in VAULT
		potential security risk with ‘RATIO_LP_USD_PER_CALL_TOK’, used in FACTORY.closeMarketForTicket
	- review interface .sol files for unused functions and delete
		verify with integration files to ensure proper public | external setting w/ modifier
	- review each external function algorithm / call stack
		- pay attention to missing/commented code that may have been lost during debugging
		- pay attention to uint type use (ie. uint32 vs 64 vs 256, etc.)
	- finalize / verify all external getter & setters needed
		- pay attention to returning structs properly (may need wrappers)
	- research & integrate ‘lock’ design pattern when needed
		maybe during DEX executions
	DONE - fix dynamic token name/symbol for new ticket creation (LIB._genTokenNameSymbol)
	DONE - integrate ability for regular CALL holders to vote at lower ratio (RATIO_CALL_TOK_PER_VOTE)
	DONE - obfuscate _ticket input for castVoteForMarketTicket (voter hash integration)
	DONE - add address markHash to MARKET_REVIEW struct in ICallitLib.sol
	DONE - migrate ‘ACCT_MARKET_REVIEWS’ (and all globals) from factory to CONFIG
	DONE - migrate DELEGATE globals over to CONFIG|M
	DONE - migrate CALL globals over to CONFIG|M
	DONE - migrate VAULT globals over to CONFIG|M
	DONE - rename CallitTicket.burnForWinLoseClaim to CallitTicket.burnForRewardClaim
	DONE - CallitFactory: fix emit event logs that were left out during compilation debugging
	DONE - CallitToken: (fwds from all contracts now)
		DONE - forward native PLS receive() over to VAULT, just like FACTORY is doing
	DONE - refactor contract addresses globals: ie. FACT_ADDR to ADDR_FACT (and for all)
	DONE - review ‘KEEPER_setContracts’ integration across all contracts
		DONE - updated to CONF_setConfig with CallitConfig.sol integration
	DONE - _uint64_from_uint256 integration review finalize
		DONE - need to review invocations w/ ERC20 based token math (balanceOf or totalSupply, etc.)
		DONE - should review all executions in general to be sure everything is cool
			DONE - reviewed OK
				- CallitFactory.sol - 2 instances OK
				- CallitLib.sol - 3 instances OK
				- CallitToken.sol - 1 instance OK
				- CallitTicket.sol - 0 instances found
			DONE - reviewed & FIX
				- CallitVault.sol - 2 instances fixed (_performTicketMint, _payPromotorDeductFeesBuyTicket)
				- CallitDelegate.sol - 1 instance fixed (closeMarketCallsForTicket)
	DONE - search and review remaining ‘LEFT OFF HERE’
	DONE - CallitVault.sol
		DONE - search & review all ‘msg.sender’ integrations (may need a ‘_receiver’ param)
	DONE - fix factory compilation size warning (DONE - latest: 27741 bytes and exceeds 24576 bytes)
		DONE - list of features temporarily removed during current debug session (need reintegration)
			DONE - 1) ERC20 integration (new CallitToken.sol contract)
			N/A  - 2) emit event logs (?)
			DONE - 3) makeNewMarket integration (migrated to DELEGATE)
	DONE - CallitDelegate.sol
		DONE - need to search & review all ‘msg.sender’ integrations (may need a ‘_receiver’ param)
	DONE - migrate / revert back to CallitLib.sol use (check compile file size)
		DONE - needs CallitVault which stores all USD stable
		DONE - factory simply convert PLS deposit and send to vault
		DONE - all deductions & payouts in external UIs (all except 2 i think),
			come from vault

latest deploys…
#-----------------------------------------------------------------------------------------------------------#
address public ADDR_LIB = address(0x84Ac3BE8Bc1393385A2ec18A4cf4A4BB93721b83); // CallitLib v0.30
address public ADDR_VAULT = address(0xE6A3288AB88B88A83D92Fb3Ba9c14699ef62b29d); // CallitVault v0.51
address public ADDR_DELEGATE = address(0xd1C73b043DF4BC75d6e210a97E9Fb6f6ef4388f7); // CallitDelegate v0.48
address public ADDR_CALL = address(0x0b8C202F4af1A7F24083e3e781fdB595464e538b); // CallitToken v0.19
address public ADDR_FACT = address(0xA80c21e68A2a2E6395Ec3f2EA486b409463ed40a); // CallitFactory v0.66
address public ADDR_CONFM = address(0xeD673eFf28fDF24ac469EC3198336C5eA5B88FFc); // CallitConfigMarket v0.4
address public ADDR_CONF = address(0x174fB6cbc83Ef578c0A36c07ac51b5a355230F43); // CallitConfig v0.21
#================================================#
#================================================#

+++ contracts/CallitConfig.sol
@@ -114,7 +114,7 @@ contract CallitConfig {
 
     // lp settings
     uint64 public MIN_USD_MARK_LIQ = 500000; // (500000 = $0.500000) min usd liquidity need for 'makeNewMarket' (total to split across all resultOptions)
-    uint16 public RATIO_LP_TOK_PER_USD = 10000; // # of ticket tokens per usd, minted for LP deploy
+    uint32 public RATIO_LP_TOK_PER_USD = 10000; // # of ticket tokens per usd, minted for LP deploy
     uint64 public RATIO_LP_USD_PER_CALL_TOK = 1000000; // (1000000 = %1.000000; 6 decimals) init LP usd amount needed per $CALL earned by market maker
         // NOTE: utilized in 'FACTORY.closeMarketForTicket'
         // LEFT OFF HERE  ... need more requirement for market maker earning $CALL
@@ -354,7 +354,7 @@ contract CallitConfig {
         require(bytes(_tok_name).length > 0 && bytes(_tok_symb).length > 0, ' invalid input  :<> ');
         CALL.setTokenNameSymbol(_tok_name, _tok_symb); // emits 'TokenNameSymbolUpdated'
     }
-    function KEEPER_setLpSettings(uint64 _usdPerCallEarned, uint16 _tokCntPerUsd, uint64 _usdMinInitLiq) external onlyKeeper {
+    function KEEPER_setLpSettings(uint64 _usdPerCallEarned, uint32 _tokCntPerUsd, uint64 _usdMinInitLiq) external onlyKeeper {
     // function KEEPER_setLpSettings(uint64 _usdPerCallEarned, uint16 _tokCntPerUsd) external onlyKeeper {
         RATIO_LP_USD_PER_CALL_TOK = _usdPerCallEarned; // LP usd amount needed per $CALL earned by market maker
         RATIO_LP_TOK_PER_USD = _tokCntPerUsd; // # of ticket tokens per usd, minted for LP deploy
